+++ src/platform/macos.rs
@@ -1,18 +1,24 @@
 use std::{env, path::PathBuf, ptr};
 
-use crate::input::KEYBOARD_LAYOUT_CHARACTER_MAP;
-
-use super::{CallbackFn, KeyModifier, KEY_DELETE, KEY_ENTER, KEY_ESCAPE, KEY_SPACE, KEY_TAB};
-use core_foundation::runloop::{kCFRunLoopCommonModes, CFRunLoop};
+mod macos_ext;
 use core_graphics::{
     event::{
-        CGEventFlags, CGEventTapLocation, CGEventTapOptions, CGEventTapPlacement, CGEventTapProxy,
-        CGEventType, CGKeyCode, EventField, KeyCode,
+        CGEventFlags, CGEventTapLocation, CGEventTapOptions, CGEventTapPlacement, CGEventType,
+        CGKeyCode, EventField, KeyCode,
     },
     sys,
 };
+pub use macos_ext::SystemTray;
 
-pub type Handle = CGEventTapProxy;
+use crate::input::KEYBOARD_LAYOUT_CHARACTER_MAP;
+use core_foundation::runloop::{kCFRunLoopCommonModes, CFRunLoop};
+
+pub use self::macos_ext::Handle;
+use self::macos_ext::{
+    new_tap, CGEventCreateKeyboardEvent, CGEventKeyboardSetUnicodeString, CGEventTapPostEvent,
+};
+
+use super::{CallbackFn, KeyModifier, KEY_DELETE, KEY_ENTER, KEY_ESCAPE, KEY_SPACE, KEY_TAB};
 
 pub const SYMBOL_SHIFT: &str = "⇧";
 pub const SYMBOL_CTRL: &str = "⌃";
@@ -79,21 +85,6 @@ fn get_char(keycode: CGKeyCode) -> Option<char> {
     None
 }
 
-#[link(name = "CoreGraphics", kind = "framework")]
-extern "C" {
-    fn CGEventTapPostEvent(proxy: CGEventTapProxy, event: sys::CGEventRef);
-    fn CGEventCreateKeyboardEvent(
-        source: sys::CGEventSourceRef,
-        keycode: CGKeyCode,
-        keydown: bool,
-    ) -> sys::CGEventRef;
-    fn CGEventKeyboardSetUnicodeString(
-        event: sys::CGEventRef,
-        length: libc::c_ulong,
-        string: *const u16,
-    );
-}
-
 pub fn send_backspace(handle: Handle, count: usize) -> Result<(), ()> {
     let null_event_source = ptr::null_mut() as *mut sys::CGEventSource;
     let (event_bs_down, event_bs_up) = unsafe {
@@ -123,121 +114,6 @@ pub fn send_string(handle: Handle, string: &str) -> Result<(), ()> {
         CGEventTapPostEvent(handle, event_str);
     }
     Ok(())
-}
-
-mod new_tap {
-    use std::{
-        mem::{self, ManuallyDrop},
-        ptr,
-    };
-
-    use core_foundation::{
-        base::TCFType,
-        mach_port::{CFMachPort, CFMachPortRef},
-    };
-    use core_graphics::{
-        event::{
-            CGEvent, CGEventMask, CGEventTapCallBackFn, CGEventTapLocation, CGEventTapOptions,
-            CGEventTapPlacement, CGEventTapProxy, CGEventType,
-        },
-        sys,
-    };
-    use foreign_types::ForeignType;
-    use libc::c_void;
-
-    type CGEventTapCallBackInternal = unsafe extern "C" fn(
-        proxy: CGEventTapProxy,
-        etype: CGEventType,
-        event: sys::CGEventRef,
-        user_info: *const c_void,
-    ) -> sys::CGEventRef;
-
-    #[link(name = "CoreGraphics", kind = "framework")]
-    extern "C" {
-        fn CGEventTapCreate(
-            tap: CGEventTapLocation,
-            place: CGEventTapPlacement,
-            options: CGEventTapOptions,
-            eventsOfInterest: CGEventMask,
-            callback: CGEventTapCallBackInternal,
-            userInfo: *const c_void,
-        ) -> CFMachPortRef;
-        fn CGEventTapEnable(tap: CFMachPortRef, enable: bool);
-    }
-
-    #[no_mangle]
-    unsafe extern "C" fn cg_new_tap_callback_internal(
-        _proxy: CGEventTapProxy,
-        _etype: CGEventType,
-        _event: sys::CGEventRef,
-        _user_info: *const c_void,
-    ) -> sys::CGEventRef {
-        let callback = _user_info as *mut CGEventTapCallBackFn;
-        let event = CGEvent::from_ptr(_event);
-        let new_event = (*callback)(_proxy, _etype, &event);
-        match new_event {
-            Some(new_event) => ManuallyDrop::new(new_event).as_ptr(),
-            None => {
-                mem::forget(event);
-                ptr::null_mut() as sys::CGEventRef
-            }
-        }
-    }
-
-    /* Generate an event mask for a single type of event. */
-    macro_rules! CGEventMaskBit {
-        ($eventType:expr) => {
-            1 << $eventType as CGEventMask
-        };
-    }
-
-    pub struct CGEventTap<'tap_life> {
-        pub mach_port: CFMachPort,
-        pub callback_ref:
-            Box<dyn Fn(CGEventTapProxy, CGEventType, &CGEvent) -> Option<CGEvent> + 'tap_life>,
-    }
-
-    impl<'tap_life> CGEventTap<'tap_life> {
-        pub fn new<F: Fn(CGEventTapProxy, CGEventType, &CGEvent) -> Option<CGEvent> + 'tap_life>(
-            tap: CGEventTapLocation,
-            place: CGEventTapPlacement,
-            options: CGEventTapOptions,
-            events_of_interest: std::vec::Vec<CGEventType>,
-            callback: F,
-        ) -> Result<CGEventTap<'tap_life>, ()> {
-            let event_mask: CGEventMask = events_of_interest
-                .iter()
-                .fold(CGEventType::Null as CGEventMask, |mask, &etype| {
-                    mask | CGEventMaskBit!(etype)
-                });
-            let cb = Box::new(Box::new(callback) as CGEventTapCallBackFn);
-            let cbr = Box::into_raw(cb);
-            unsafe {
-                let event_tap_ref = CGEventTapCreate(
-                    tap,
-                    place,
-                    options,
-                    event_mask,
-                    cg_new_tap_callback_internal,
-                    cbr as *const c_void,
-                );
-
-                if !event_tap_ref.is_null() {
-                    Ok(Self {
-                        mach_port: (CFMachPort::wrap_under_create_rule(event_tap_ref)),
-                        callback_ref: Box::from_raw(cbr),
-                    })
-                } else {
-                    _ = Box::from_raw(cbr);
-                    Err(())
-                }
-            }
-        }
-
-        pub fn enable(&self) {
-            unsafe { CGEventTapEnable(self.mach_port.as_concrete_TypeRef(), true) }
-        }
-    }
 }
 
 pub fn run_event_listener(callback: &CallbackFn) {
